+++ languagetool-core/src/main/java/org/languagetool/rules/RuleMatch.java
@@ -373,8 +373,12 @@ public class RuleMatch implements Comparable<RuleMatch> {
   @Override
   public String toString() {
     if (rule instanceof PatternRule) {
+      //String covered = getSentence().getText().substring(getFromPos(), getToPos());
+      //return ((PatternRule) rule).getFullId() + ":" + offsetPosition + ":" + message + ":" + covered + " -> " + getSuggestedReplacements();
       return ((PatternRule) rule).getFullId() + ":" + offsetPosition + ":" + message;
     } else {
+      //String covered = getSentence().getText().substring(getFromPos(), getToPos());
+      //return rule.getId() + ":" + offsetPosition + ":" + message + ":" + covered + " -> " + getSuggestedReplacements();
       return rule.getId() + ":" + offsetPosition + ":" + message;
     }
   }