+++ languagetool-standalone/CHANGES.md
@@ -2,9 +2,12 @@
 
 ## 4.7 (release planned for 2019-09-26)
 
-...
-
-
+#### General
+  * Spell suggestion improvements: for many cases of a misplaced space,
+    the suggestion are now better. For example, "thef eedback" can now
+    be corrected to "the feedback" in one step. (#1729)
+  
+  
 
 ## 4.6 (2019-06-26)